mv toSVG with <base> test to toImage suite
